Move catch-all route to ServiceProviderInterface::boot()
This way it gets registered after the dynamic OPTIONS routes that get registered by CorsServiceProvider::boot()
